Added simple tests

+++ tests/input/div_zero.c
@@ -1,2 +1,2 @@
 // Test for a division by zero
-double y = 5/0;
+float y = 5/0;
